chore: split out gatewayapiconformance tests pkg
allow pilot gateway-api conformance tests to be called independently
